Heavy ion Vs algorithm same event subtraction
