+++ RecoHI/HiJetAlgos/plugins/VoronoiBackgroundProducer.cc
@@ -28,6 +28,7 @@
 #include "JetMETCorrections/Objects/interface/JetCorrectionsRecord.h"
 
 #include "VoronoiAlgorithm.h"
+#include "SameEventVoronoiAlgorithm.h"
 
 using namespace std;
 //
@@ -47,11 +48,20 @@ class VoronoiBackgroundProducer : public edm::EDProducer {
       // ----------member data ---------------------------
 
    edm::EDGetTokenT<reco::CandidateView> src_;
-   VoronoiAlgorithm* voronoi_;
+   unique_ptr<GenericVoronoiAlgorithm> voronoi_;
    bool doEqualize_;
    double equalizeThreshold0_;
    double equalizeThreshold1_;
    double equalizeR_;
+   bool excludeV1_;
+   int maxVn_;
+   bool diagonalVn_;
+   bool sameEvent_;
+   double sameEventAntiktDistance_;
+   double sameEventExclusionPtMin_;
+   double sameEventExclusionRadius_;
+   double sameEventFakeRejectEtMax_;
+   double sameEventFakeRejectEtMaxOverMean_;
    bool useTextTable_;
    bool jetCorrectorFormat_;
    bool isCalo_;
@@ -75,11 +85,19 @@ class VoronoiBackgroundProducer : public edm::EDProducer {
 // constructors and destructor
 //
 VoronoiBackgroundProducer::VoronoiBackgroundProducer(const edm::ParameterSet& iConfig):
-   voronoi_(0),
    doEqualize_(iConfig.getParameter<bool>("doEqualize")),
    equalizeThreshold0_(iConfig.getParameter<double>("equalizeThreshold0")),
    equalizeThreshold1_(iConfig.getParameter<double>("equalizeThreshold1")),
    equalizeR_(iConfig.getParameter<double>("equalizeR")),
+   excludeV1_(iConfig.getParameter<bool>("excludeV1")),
+   maxVn_(iConfig.getParameter<int>("maxVn")),
+   diagonalVn_(iConfig.getParameter<bool>("diagonalVn")),
+   sameEvent_(iConfig.getParameter<bool>("sameEvent")),
+   sameEventAntiktDistance_(iConfig.getParameter<double>("sameEventAntiktDistance")),
+   sameEventExclusionPtMin_(iConfig.getParameter<double>("sameEventExclusionPtMin")),
+   sameEventExclusionRadius_(iConfig.getParameter<double>("sameEventExclusionRadius")),
+   sameEventFakeRejectEtMax_(iConfig.getParameter<double>("sameEventFakeRejectEtMax")),
+   sameEventFakeRejectEtMaxOverMean_(iConfig.getParameter<double>("sameEventFakeRejectEtMaxOverMean")),
    useTextTable_(iConfig.getParameter<bool>("useTextTable")),
    jetCorrectorFormat_(iConfig.getParameter<bool>("jetCorrectorFormat")),
    isCalo_(iConfig.getParameter<bool>("isCalo")),
@@ -122,7 +140,7 @@ VoronoiBackgroundProducer::produce(edm::Event& iEvent, const edm::EventSetup& iS
         if(isData) calibrationFile = "RecoHI/HiJetAlgos/data/ue_calibrations_hermite_pf_data.txt";
         if(!isData) calibrationFile = "RecoHI/HiJetAlgos/data/ue_calibrations_hermite_pf_mc.txt";
       }
-	ue = new UECalibration(calibrationFile);
+	  ue = new UECalibration(calibrationFile);
 	}
 	else if (jetCorrectorFormat_) {
 		edm::ESHandle<JetCorrectorParametersCollection> ueHandle;
@@ -143,7 +161,12 @@ VoronoiBackgroundProducer::produce(edm::Event& iEvent, const edm::EventSetup& iS
 	 ue = new UECalibration(ueTable->values);
 	}
 
-     voronoi_ = new VoronoiAlgorithm(ue,equalizeR_,std::pair<double, double>(equalizeThreshold0_,equalizeThreshold1_),doEqualize_);
+	if (sameEvent_) {
+		voronoi_ = make_unique<SameEventVoronoiAlgorithm>(sameEventAntiktDistance_, sameEventExclusionPtMin_, sameEventExclusionRadius_, sameEventFakeRejectEtMax_, sameEventFakeRejectEtMaxOverMean_, std::pair<double, double>(equalizeThreshold0_,equalizeThreshold1_), doEqualize_);
+	}
+	else {
+		voronoi_ = make_unique<VoronoiAlgorithm>(ue,equalizeR_,excludeV1_,maxVn_,diagonalVn_,std::pair<double, double>(equalizeThreshold0_,equalizeThreshold1_),doEqualize_);
+	}
    }
 
    voronoi_->clear();
