+++ RecoHI/HiJetAlgos/python/HiRecoJets_cff.py
@@ -29,6 +29,15 @@ voronoiBackgroundCalo = cms.EDProducer('VoronoiBackgroundProducer',
                                        equalizeThreshold0 = cms.double(5.0),
                                        equalizeThreshold1 = cms.double(35.0),
                                        equalizeR = cms.double(0.4),
+                                       excludeV1 = cms.bool(True),
+                                       maxVn = cms.int32(2),
+                                       diagonalVn = cms.bool(True),
+                                       sameEvent = cms.bool(False),
+                                       sameEventAntiktDistance = cms.double(0.2),
+                                       sameEventExclusionPtMin = cms.double(25.0),
+                                       sameEventExclusionRadius = cms.double(0.4),
+                                       sameEventFakeRejectEtMax = cms.double(3),
+                                       sameEventFakeRejectEtMaxOverMean = cms.double(4),
 				       useTextTable = cms.bool(False),
 				       jetCorrectorFormat = cms.bool(True),
                                        isCalo = cms.bool(True),
